fix comments from CR
