+++ src/backend/api/Fusion.Resources.Logic/Requests/Authorization/CanApproveStepHandlerBase.cs
@@ -1,5 +1,4 @@
-﻿using Fusion.AspNetCore.FluentAuthorization;
-using Fusion.Authorization;
+﻿using Fusion.Authorization;
 using Fusion.Resources.Api.Authorization.Requirements;
 using Fusion.Resources.Authorization.Requirements;
 using Fusion.Resources.Database.Entities;
@@ -10,28 +9,6 @@ using System.Threading.Tasks;
 
 namespace Fusion.Resources.Logic.Requests
 {
-    public record WorkflowAccessKey(string Subtype, string CurrentStep)
-    {
-        public static implicit operator WorkflowAccessKey((string, string) tuple)
-            => new WorkflowAccessKey(tuple.Item1, tuple.Item2);
-    }
-    public record WorkflowAccess()
-    {
-        public bool IsResourceOwnerAllowed { get; init; } = false;
-        public bool IsParentResourceOwnerAllowed { get; init; } = false;
-        public bool IsSiblingResourceOwnerAllowed { get; init; } = false;
-        public bool IsAllResourceOwnersAllowed { get; init; } = false;
-        public bool IsCreatorAllowed { get; init; } = false;
-        public bool IsDirectTaskOwnerAllowed { get; init; } = false;
-        public bool IsOrgChartTaskOwnerAllowed { get; init; } = false;
-        public bool IsOtherProjectMembersAllowed { get; init; } = false;
-        public bool IsOrgChartReadAllowed { get; init; } = false;
-        public bool IsOrgChartWriteAllowed { get; init; } = false;
-        public bool IsOrgAdminAllowed { get; init; } = false;
-
-        public static WorkflowAccess Default = new WorkflowAccess();
-    }
-
     public class CanApproveStepHandlerBase
     {
         private readonly IAuthorizationService authorizationService;
