+++ src/backend/api/Fusion.Resources.Logic/Requests/Authorization/WorkflowAccessKey.cs
@@ -1,0 +1,8 @@
+﻿namespace Fusion.Resources.Logic.Requests
+{
+    public record WorkflowAccessKey(string Subtype, string CurrentStep)
+    {
+        public static implicit operator WorkflowAccessKey((string, string) tuple)
+            => new WorkflowAccessKey(tuple.Item1, tuple.Item2);
+    }
+}
